moved routes.js to src/app, updated routes imports
