+++ src/Navigation/Header/index.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import { toMovieList, toPeopleList } from "../../routes";
+import { toMovieList, toPeopleList } from "../../app/routes";
 import { StyledHeader, Logo, Title, StyledList, StyledNavLink } from "./styled";
 export const Header = () => (
     <StyledHeader>
